Add a solution directory (package) to each exercise module - solution code will compile now and easier to view the full solution. Also needed to make the TopicLoader methods public
