+++ kafka-streams/src/main/java/io/confluent/developer/aggregate/solution/StreamsAggregate.java
@@ -1,6 +1,7 @@
-package io.confluent.developer.aggregate;
+package io.confluent.developer.aggregate.solution;
 
 import io.confluent.developer.StreamsUtils;
+import io.confluent.developer.aggregate.TopicLoader;
 import io.confluent.developer.avro.ElectronicOrder;
 import io.confluent.kafka.streams.serdes.avro.SpecificAvroSerde;
 import org.apache.kafka.common.serialization.Serdes;
