+++ kafka-streams/src/main/java/io/confluent/developer/basic/solution/BasicStreams.java
@@ -1,5 +1,6 @@
-package io.confluent.developer.serdes;
+package io.confluent.developer.basic.solution;
 
+import io.confluent.developer.basic.TopicLoader;
 import org.apache.kafka.common.serialization.Serdes;
 import org.apache.kafka.streams.KafkaStreams;
 import org.apache.kafka.streams.StreamsBuilder;
@@ -12,11 +13,13 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.util.Properties;
 
-public class StreamsSerdes {
+public class BasicStreams {
 
     public static void main(String[] args) throws IOException {
-        final Properties streamsProps = new Properties();
-        streamsProps.load(new FileInputStream("src/main/resources/streams.properties"));
+        Properties streamsProps = new Properties();
+        try (FileInputStream fis = new FileInputStream("src/main/resources/streams.properties")) {
+            streamsProps.load(fis);
+        }
         streamsProps.put(StreamsConfig.APPLICATION_ID_CONFIG, "basic-streams");
 
         StreamsBuilder builder = new StreamsBuilder();
@@ -26,14 +29,17 @@ public class StreamsSerdes {
         final String orderNumberStart = "orderNumber-";
         KStream<String, String> firstStream = builder.stream(inputTopic, Consumed.with(Serdes.String(), Serdes.String()));
 
-        firstStream.filter((key, value) -> value.contains(orderNumberStart))
-                .mapValues(value -> value.substring(value.indexOf(orderNumberStart)))
-                .filter((key, value) -> Long.parseLong(value) > 1000)
-                .to(outputTopic, Produced.with(Serdes.String(), Serdes.String()));
+        firstStream.peek((key, value) -> System.out.println("Incoming record - key " +key +" value " + value))
+                   .filter((key, value) -> value.contains(orderNumberStart))
+                   .mapValues(value -> value.substring(value.indexOf("-") + 1))
+                   .filter((key, value) -> Long.parseLong(value) > 1000)
+                   .peek((key, value) -> System.out.println("Outgoing record - key " +key +" value " + value))
+                   .to(outputTopic, Produced.with(Serdes.String(), Serdes.String()));
 
         KafkaStreams kafkaStreams = new KafkaStreams(builder.build(), streamsProps);
+        
+        TopicLoader.runProducer();
+
         kafkaStreams.start();
     }
-
-
 }
