+++ kafka-streams/src/main/java/io/confluent/developer/ktable/solution/KTableExample.java
@@ -1,5 +1,6 @@
-package io.confluent.developer.ktable;
+package io.confluent.developer.ktable.solution;
 
+import io.confluent.developer.ktable.TopicLoader;
 import org.apache.kafka.common.serialization.Serdes;
 import org.apache.kafka.common.utils.Bytes;
 import org.apache.kafka.streams.KafkaStreams;
