+++ kafka-streams/src/main/java/io/confluent/developer/processor/solution/ProcessorApi.java
@@ -1,6 +1,7 @@
-package io.confluent.developer.processor;
+package io.confluent.developer.processor.solution;
 
 import io.confluent.developer.avro.ElectronicOrder;
+import io.confluent.developer.processor.TopicLoader;
 import io.confluent.kafka.streams.serdes.avro.SpecificAvroSerde;
 import org.apache.kafka.common.serialization.Serde;
 import org.apache.kafka.common.serialization.Serdes;
